Tiny cosmetic fix to some types

+++ ai/mmcheat.h
@@ -8,12 +8,12 @@
 #define AI_MMCHEAT_H
 
 /* How many plies we search before evaluating */
-static const uint_fast8_t MaxPlies = 7;
+static const size_t MaxPlies = 7;
 
 /* The hand size at which we cease considering drawing more cards. Letting this
  * be unbounded results in the AI drawing so many cards that its evaluation to
  * MaxPlies plies becomes too slow. */
-static const uint_fast8_t MaxCards = CPP + 3;
+static const size_t MaxCards = CPP + 3;
 
 uint_fast32_t aiMmCheat(const struct aistate* const restrict as)
 	__attribute__((hot,pure,nonnull));
